fix: Add support for nested configs
