+++ test/src/settings.test.js
@@ -17,7 +17,6 @@ test('settings: default settings for falsy config', function (t) {
   t.deepEqual(pluginSettings.paths, {});
   t.deepEqual(pluginSettings.allExceptPaths, {});
   t.deepEqual(pluginSettings.pathsOptions, settings.defaults.pathsOptions);
-  t.deepEqual(pluginSettings.extensions, settings.defaults.extensions);
   t.deepEqual(pluginSettings.plugins, []);
   t.end();
 });
@@ -28,7 +27,6 @@ test('settings: default settings for undefined settings', function (t) {
   t.deepEqual(pluginSettings.paths, {});
   t.deepEqual(pluginSettings.allExceptPaths, {});
   t.deepEqual(pluginSettings.pathsOptions, settings.defaults.pathsOptions);
-  t.deepEqual(pluginSettings.extensions, settings.defaults.extensions);
   t.deepEqual(pluginSettings.plugins, []);
   t.end();
 });
@@ -40,7 +38,6 @@ test('settings: default settings for undefined plugin settings', function (t) {
   t.deepEqual(pluginSettings.paths, {});
   t.deepEqual(pluginSettings.allExceptPaths, {});
   t.deepEqual(pluginSettings.pathsOptions, settings.defaults.pathsOptions);
-  t.deepEqual(pluginSettings.extensions, settings.defaults.extensions);
   t.deepEqual(pluginSettings.plugins, []);
   t.end();
 });
@@ -54,20 +51,7 @@ test('settings: default settings for empty plugin settings', function (t) {
   t.deepEqual(pluginSettings.paths, {});
   t.deepEqual(pluginSettings.allExceptPaths, {});
   t.deepEqual(pluginSettings.pathsOptions, settings.defaults.pathsOptions);
-  t.deepEqual(pluginSettings.extensions, settings.defaults.extensions);
   t.deepEqual(pluginSettings.plugins, []);
-  t.end();
-});
-
-test('settings: use passed extensions', function (t) {
-  var pluginSettings = settings.prepare({
-    settings: {
-      'eslint-plugin-disable': {
-        extensions: ['.foo', '.bar']
-      }
-    }
-  });
-  t.deepEqual(pluginSettings.extensions, ['.foo', '.bar']);
   t.end();
 });
 
@@ -161,18 +145,6 @@ test('settingsL convert allExceptPaths to arrays', function (t) {
     'foo': ['/foo/1'],
     'bar': ['/bar/1']
   });
-  t.end();
-});
-
-test('settings: convert extensions to array', function (t) {
-  var pluginSettings = settings.prepare({
-    settings: {
-      'eslint-plugin-disable': {
-        extensions: '.foo'
-      }
-    }
-  });
-  t.deepEqual(pluginSettings.extensions, ['.foo']);
   t.end();
 });
